refactor: 디렉토리 복수형으로 수정
