+++ RTWTest/Webserver/Authentication/JwtTokenProviderTests.cs
@@ -1,5 +1,5 @@
 using Microsoft.Extensions.Configuration;
-using RTWWebServer.Provider.Authentication;
+using RTWWebServer.Providers.Authentication;
 
 namespace RTWTest.Webserver.Authentication;
