--- RTWWebServer/DTO/Request/CreateGuestRequest.cs
@@ -1,5 +1,0 @@
-namespace RTWWebServer.DTO.Request;
-
-public class CreateGuestRequest
-{
-}
