+++ RTWWebServer/Entities/UserSession.cs
@@ -1,4 +1,4 @@
-namespace RTWWebServer.Entity;
+namespace RTWWebServer.Entities;
 
 public class UserSession(int userId, string token)
 {
