+++ RTWWebServer/Services/AccountService.cs
@@ -1,9 +1,9 @@
 using NetworkDefinition.ErrorCode;
 using RTWWebServer.Exceptions;
-using RTWWebServer.Provider.Authentication;
-using RTWWebServer.Repository;
+using RTWWebServer.Providers.Authentication;
+using RTWWebServer.Repositories;
 
-namespace RTWWebServer.Service;
+namespace RTWWebServer.Services;
 
 public class AccountService(
     IUnitOfWork unitOfWork,
